🎨 Rename folder and files to Empty Activity
