+++ src/app/components/EmptyActivity/EmptyActivity.module.css
